Extract players, teams information from cricinfo website
This will get all the players and their details from the cricinfo
website for a given BBL(Big Bash League) cricket team. The details can
then be serialized to objects and then store them using entity
framework.

This is done using HttpClient and the third party tool Html Agilitypack
for .Net and Fizzler.